Moved components, Renamed props
