+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/Card.module.scss
