+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/index.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from 'react'
-import { Card } from './Card/Card'
+import { Card } from './Card'
 import { IProjectCardProps } from './types'
 
 export const ProjectCard: FunctionComponent<IProjectCardProps> = (props) => {
